requirements.txt
# Core dependencies
playwright==1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Scheduling
apscheduler>=3.10.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Configuration
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging and utils
loguru>=0.7.0
python-dotenv>=1.0.0
tenacity>=8.2.0

# Image processing
pillow>=10.1.0
